Automation/requirements.txt
redis>=4.0.0
zstd>=1.5.0
watchdog>=2.0.0
pathspec>=0.9.0
python-dotenv>=1.0.0